requirements.txt
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext[openai]==0.5.7
python-dotenv>=1.0.0
